web/requirements.txt
django>=4.0.0,<4.1.0
djangorestframework
django-rest-passwordreset
pyyaml
requests
django-cors-headers
django-rest-knox
Pillow
gunicorn
python-dotenv
psycopg2-binary
django-storages
boto3
